pyproject.toml
[tool.poetry]
name = "bullyguard"
version = "0.1.0"
description = ""
authors = ["BarG3126 <bar3126@gmail.com>"]
readme = "README.md"
packages = [
    { include = "bullyguard" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pydantic = "~=2.10.5"
hydra-core = "~=1.3.2"
pandas = "~=2.0"
fsspec = {version = "2024.12.0", extras = ["gcs"]}
gcsfs = "=2024.12.0"
transformers = {version = "~=4.49.0", extras = ["torch"]}
torch = "~=2.6.0"
lightning = "~=2.5.0"
torch-metrics = "~=1.1.7"
mlflow ="~=2.20.2"
psycopg2-binary = "~=2.9.6"
google-cloud-compute = "~=1.26.0"
matplotlib = "~=3.10.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "4.3.5"
pytest = "~7.4"
black = "~23.7.0"
isort = "~5.12"
flake8 = "~6.0"
mypy = "~1.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["bullyguard"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true
warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true
verbosity = 0